+++ src/main/java/ru/accumulator/MerkleTree.java
@@ -19,7 +19,7 @@ public class MerkleTree {
         }
 
         original_data = new ArrayList<>(source);
-        int size = (int) AccumulatorUtils.max_leq_pow2(4 * source.size() - 1);
+        int size =  AccumulatorUtils.max_leq_pow2(4 * source.size() - 1);
         //assert (size/2 >= source.size());
         capacity = source.size();
         size--;
@@ -49,7 +49,7 @@ public class MerkleTree {
     private void expand() {
         List<byte[]> new_storage = new ArrayList<>(Collections.nCopies(2 * hashed_data.size() + 1, null));
         for (int i = 0; i < hashed_data.size(); i++) {
-            new_storage.set(i + (int) AccumulatorUtils.max_leq_pow2((long) i + 1), hashed_data.get(i));
+            new_storage.set(i +  AccumulatorUtils.max_leq_pow2( i + 1), hashed_data.get(i));
         }
 
         new_storage.set(0, getVertexesHash(0));
@@ -121,7 +121,7 @@ public class MerkleTree {
         return false;
     }*/
 
-    public static boolean verify(byte[] rootHash, long index, byte[] leafHash, List<byte[]> neighboursHashes) {
+    public static boolean verify(byte[] rootHash, int index, byte[] leafHash, List<byte[]> neighboursHashes) {
         byte[] current = leafHash;
         for (byte[] neighbourString : neighboursHashes) {
             if (index % 2 == 0) {
